+++ errors/errors_test.go
@@ -70,7 +70,8 @@ func (s *ErrorsSuite) TestPositive(c *C) {
 		},
 	)
 	c.Assert(errs.Add(nil), NotNil)
-	c.Assert(errs.Error("  "), Equals, "  1\n  2\n  3\n  4\n  5")
+	c.Assert(errs.ErrorWithPrefix("  "), Equals, "  1\n  2\n  3\n  4\n  5")
+	c.Assert(errs.Error(), Equals, "1\n2\n3\n4\n5")
 
 	errs.Reset()
 
@@ -142,7 +143,8 @@ func (s *ErrorsSuite) TestNegative(c *C) {
 	c.Assert(errs.All(), HasLen, 0)
 	c.Assert(errs.IsEmpty(), Equals, true)
 	c.Assert(errs.Last(), IsNil)
-	c.Assert(errs.Error(""), Equals, "")
+	c.Assert(errs.Error(), Equals, "")
+	c.Assert(errs.ErrorWithPrefix(""), Equals, "")
 }
 
 func (s *ErrorsSuite) TestNil(c *C) {
@@ -154,7 +156,8 @@ func (s *ErrorsSuite) TestNil(c *C) {
 	c.Assert(errs.IsEmpty(), Equals, true)
 	c.Assert(errs.First(), IsNil)
 	c.Assert(errs.Last(), IsNil)
-	c.Assert(errs.Error(""), Equals, "")
+	c.Assert(errs.Error(), Equals, "")
+	c.Assert(errs.ErrorWithPrefix(""), Equals, "")
 	c.Assert(errs.Get(10), IsNil)
 }
 
@@ -167,7 +170,8 @@ func (s *ErrorsSuite) TestNoInit(c *C) {
 	c.Assert(errs.IsEmpty(), Equals, true)
 	c.Assert(errs.First(), IsNil)
 	c.Assert(errs.Last(), IsNil)
-	c.Assert(errs.Error(""), Equals, "")
+	c.Assert(errs.Error(), Equals, "")
+	c.Assert(errs.ErrorWithPrefix(""), Equals, "")
 
 	c.Assert(errs.Add(errors.New("1")), NotNil)
 	c.Assert(errs.Last(), DeepEquals, errors.New("1"))